Revert "refactor: move BaseMethodsFacade in facades"
This reverts commit ad884f9a01223426bb23f4b666929bab77533397.
